Move PeopleMover.tsx and ErrorPageTemplate.tsx files into their own folder
