+++ ui/src/ErrorPageTemplate/ErrorPageTemplate.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../Styles/Colors';
 
 .ErrorPageContainer {
   text-align: center;
