+++ ui/src/Routes.tsx
@@ -24,8 +24,8 @@ import {AuthenticatedRoute} from './Auth/AuthenticatedRoute';
 import RedirectWrapper from './ReusableComponents/RedirectWrapper';
 import SpaceDashboard from './SpaceDashboard/SpaceDashboard';
 import AuthorizedRoute from './Auth/AuthorizedRoute';
-import PeopleMover from './Application/PeopleMover';
-import ErrorPageTemplate from './Application/ErrorPageTemplate';
+import PeopleMover from './PeopleMover/PeopleMover';
+import ErrorPageTemplate from './ErrorPageTemplate/ErrorPageTemplate';
 import TimeOnProduct from './TimeOnProductPage/TimeOnProduct';
 import AnimatedImageSrc from './Assets/404.gif';
 import errorImageSrc from './Assets/403.png';
