+++ ui/src/tests/PeopleMover.test.tsx
@@ -17,7 +17,7 @@
 
 import React from 'react';
 import TestUtils, {getApplicationSetup, renderWithRedux} from './TestUtils';
-import PeopleMover from '../Application/PeopleMover';
+import PeopleMover from '../PeopleMover/PeopleMover';
 import {screen, waitFor} from '@testing-library/react';
 import {Router} from 'react-router-dom';
 import {createBrowserHistory, History, Location} from 'history';
